+++ src/components/Breadcrumb/Divider/Divider.css
@@ -1,0 +1,3 @@
+.arrow {
+  @apply shrink-0 mx-2 size-4 text-black;
+}
