+++ src/components/ButtonsWrapper/ButtonsWrapper.css
@@ -1,0 +1,12 @@
+#group-buttons {
+  width: 100%;
+
+  display: flex;
+  justify-content: center;
+
+  margin-top: 32px;
+}
+
+.group-buttons {
+  @apply flex flex-col items-center justify-center md:flex-row md:justify-start space-y-6 md:space-y-0 md:space-x-8;
+}
