+++ src/components/Illustration/Illustration.css
@@ -1,0 +1,4 @@
+#logo {
+  width: 100%;
+  max-width: 600px;
+}
