+++ src/components/Illustration/Illustration.jsx
@@ -1,0 +1,12 @@
+import poemixLogo from "../../assets/homepage/poemix_logo.webp";
+import "./Illustration.css";
+
+function Illustration() {
+  return (
+    <>
+      <img src={poemixLogo} id="logo" alt="Poemix logo" />
+    </>
+  );
+}
+
+export default Illustration;
