+++ src/pages/Home/Home.css
@@ -12,7 +12,7 @@ html {
   user-select: none; /* Non-prefixed version, currently */
 }
 
-#app {
+#wrapper {
   height: 100vh;
 
   padding: 0 30px;
@@ -23,22 +23,4 @@ html {
   justify-content: center;
   align-items: center;
   flex-direction: column;
-}
-
-#logo {
-  width: 100%;
-  max-width: 600px;
-}
-
-#group-buttons {
-  width: 100%;
-
-  display: flex;
-  justify-content: center;
-
-  margin-top: 32px;
-}
-
-.group-buttons {
-  @apply flex flex-col items-center justify-center md:flex-row md:justify-start space-y-6 md:space-y-0 md:space-x-8;
 }
