+++ src/pages/Zines/Zines.css
@@ -1,0 +1,18 @@
+.publication-listing {
+  padding: 0 30px;
+  margin-top: 30px;
+
+  overflow: hidden;
+
+  display: flex;
+  justify-content: center;
+  flex-direction: column;
+}
+
+.title {
+  text-align: center;
+}
+
+.link {
+  text-align: center;
+}
